Add a formal proof for some of BytesAndTimestamps properties
This adds a automated formal proof for some of the properties of
BytesAndTimestamps. It uses Kani model checker, which transforms rust's
MIR into SAT/SMT and does exhaustive search (proof).

The BytesAndTimestamps are extracted from telio-wg, because Kani doesn't
support SSE(2) but telio-wg depends on ring, which assumes SSE(2) for
intel targets.

Custom Instant is used in the Kani build, because Kani doesn't support
required syscalls.
