+++ crates/telio-wg/Cargo.toml
@@ -42,7 +42,6 @@ itertools.workspace = true
 mockall.workspace = true
 ntest.workspace = true
 pretty_assertions.workspace = true
-rstest.workspace = true
 tokio = { workspace = true, features = ["test-util"] }
 
 telio-firewall.workspace = true
